Clean Up
Remove parts of code that are not needed. Added code for the new textures so it now follows whatever Ls: toasts says the colour for the border should be. No longer includes Icon borders as this is part of the toast border to make it look cleaner.

Anything that would have a white border (currency etc) now defaults to black border, You shouldn't see any changes from what it looks like now.